src/requirements.txt
discord.py
pytest
urllib3[secure]
webhooks
sentry-sdk
